RR-647 Restructured tests
